Ops, RTL wasn't correctly updated.
git-svn-id: svn://falconpl.org/falcon/core/trunk@442 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ rtl/cmdlineparser.cpp
@@ -287,7 +287,7 @@ FALCON_FUNC  CmdlineParser_parse( ::Falcon::VMachine *vm )
          else
          {
             vm->retval( false );
-            self->setProperty( "lastParsed", (int64) i );
+            self->setProperty( "lastParsed", i );
             return;
          }
       }
@@ -307,7 +307,7 @@ FALCON_FUNC  CmdlineParser_parse( ::Falcon::VMachine *vm )
          else
          {
             vm->retval( false );
-            self->setProperty( "lastParsed", (int64) i );
+            self->setProperty( "lastParsed", i );
             return;
          }
       }
@@ -344,7 +344,7 @@ FALCON_FUNC  CmdlineParser_parse( ::Falcon::VMachine *vm )
             else
             {
                vm->retval( false );
-               self->setProperty( "lastParsed", (int64) i );
+               self->setProperty( "lastParsed", i );
                return;
             }
          }
